requirements.txt
fastapi
uvicorn[standard]
firebase-admin
requests
beautifulsoup4
python-dotenv